deploy/requirements.txt
streamlit
tensorflow
matplotlib
seaborn
scikit-learn
tqdm
pillow
onnxruntime
tf2onnx